Better mentoring (#2272)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Be voice pref saves like it should.

Mentor helps play sounds for sending and receiving but they are not
unique at this time.

Updated css for mentor helps:
Admins:

![image](https://github.com/user-attachments/assets/8aacba51-6fcf-418e-ac8e-5ba4e91c2ddc)
Players:

![image](https://github.com/user-attachments/assets/ac78a5af-f670-48eb-9bf2-eccfb0cb86f8)
Timeout:

![image](https://github.com/user-attachments/assets/38b302de-6887-4297-90b6-3ba2a574b1d1)

Actual ADMIN LOG: message for mentor help answering so its visible, this
may be annoying and the prefs don't work(?) so sorry admins.

Admin jumps on mentor helps don't both lead to the subject of the mentor
help.

closes #668
closes #1505
closes #1506

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Mentors and players get to mentor better

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->