FIXES !! (crafting prices , Fluff structures , Apothecary machinery , and more!) (#2330)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->
-looking on the telescope can give you nocs blessing or the wrath of
astrata.
-arachnic statue mini quest for Dark elfs. Bring her Ears and she will
reward you with a Triumph. ( only works once)
-Crafted items will now have standard price unless stated otherwise on
its respective recipe.
-Leather coat is now craftable
-Essence Splitter can now eject its contents by right clicking it.
-Essence Infuser now crafts the correct type of Primordial crystal , no
more invisible shards.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->
real tiny things but should give some nice QoL overall.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server. (super minor changes)
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->