Alchemy and Magic Tweaks (#2351)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Manabloom flowers can now be grinded with a pestle.
Removed "choice" menu from the mortar in favor of the following
behaviour , ( left click = alchemy , righ click = grind/juice) this is
because it was insanely anoying for apothecaries to grind multiple
products with a menu popping over and over.
Removed Manabloon Powder "crafting" recipe.
Manablooms now have a chance to give 1 extra dust and 1 extra runedust
when grinded.
Increased fish water dust Yield.
Increased Fyritus fire dust Yield.


## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
Fixed some small anoyances of alchemy and mages. 
Adds a way to obtain runedust.
makes fish and firitus farming a little more rewarding considering they
are not readily available.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->