Fixes Thieves' Cant, Adds More Verbs to Thieves' Cant. People Without Tongues Don't Make Strange Sounds Anymore (Garbles Speech). (#2338)
## About The Pull Request
Thieves' Cant "should" work like it is supposed to now. People who
understand it have language icons and undisguised text so they can know
and understand what the person speaking it says. You used to need to not
have a tongue to speak theives' cant, now you can speak it with a
tongue, and if you don't have a tongue your language isn't garbled while
speaking thieves cant. People in sight of someone speaking thieves' cant
just see a random emote.

Fixes tongueless speech, by giving the flag to every playable sentient
language besides celestial (I thought the language of Gods is so fancy
it is un-recognizable if spoken without a tongue). Now instead of
"Tongueless Nerd makes a strange sound." you'll get what they typed with
tongue needed letters being changed to something else.

## Why It's Good For The Game
Thieves can now actually coordinate stealthily, and people without
tongues aren't just mute.

Can't really provide pictures for this one because of how it was tested,
but I had my brother log in and test it on my private server. It didn't
runtime, didn't spill out past view range, you can't yell with it to go
farther, and it doesn't get picked up by SCOMs. This is going to need a
test. As I've said before I'm not smart enough to fix thieves cant,
luckily this guy
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/2579/files
is. This is going to need a review though, since again, I'm not that
smart with this kind of stuff, and there is a chance I messed something
up, or pulled something totally unnecessary.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->