Prevent server crashes by cleaning up HTTP requests sent by SSelastic and Plexora (#2379)
## About The Pull Request

okay so the logs from
https://github.com/Monkestation/Vanderlin/pull/2374 made it clear what
the issue was - SSplexora and SSelastic sent a lot of http requests,
_which were never properly cleaned up_ due to how rust-g works, and they
could eventually OOM BYOND (which is 32-bit) if they stacked for too
long. so. those 2 subsystems properly clean up after themselves now.

## Why It's Good For The Game

bugfix good

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this ~~on a local server.~~ in production
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->