Sound cache for ambience (RUSTG) + kill SSdroning in favour of SEND_SOUND and channels (#2269)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Caches sound lengths for ambient track cooldowns, updates Rust-G to
3.9.0 for this.

Tears out SSdroning to use the TG ambient sound system adjusted for us
instead.

This should fix the issue of combat mode music failing to stop or
breaking entirely.

Area stings use SEND_SOUND.

Spooky ambience will play if the tile you are on is dark, otherwise
normal ambience will play depending on time of day, being in combat mode
overrides background ambience and so does spice and being a maniac but
combat mode takes precendece over those two.

How the old droning system handled this is poor as "ambientsounds" is
actually an alt droning track played in the background (most of the time
inaudible) and "spookysounds" is actually ambient sounds.

Many of the ambient sounds (NOT spooky) themselves as files are too low
a volume to be heard even without the droning tracks, even at 100%
volume being sent directly.

Downsides of not using SSdroning:
Even if it was broken sometimes and mostly superfluous it did give a
finer control over these looping sounds, if such a thing is needed then
it can easily happen in SSambience instead.

For the end player there should be hardly any difference besides from
ambient sounds being more frequent and probably louder. The only
difference being the lack of a background track in darkness for the
SPOOKY droning instead.

TODO?
- [x] Alt droning tracks to replace the buzz, say the lights go out or
you go into a dark alley? now you get a spooky droner. Or more
accurately disable the background music and just leave the ambient
droner.
- [ ] Look over some of the ambient tracks if its easy to just bump up
the volume

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
It should just be

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->