Improved reagent handling during crafting (#2380)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

The PR changes the way reagents are deleted during the crafting process
to fix the issue with reagents being deleted even on crafting failure
and an infinite crafting loop that occurs as you continue trying to
craft something using a now empty reagent container.

Instead of deletion, it now transfers reagents into a generic holder in
nullspace, one for each bottle used. On failure, reagents are
transferred back to their associated bottle and the generic holders are
nuked from existence, while on success the holders are nuked but
reagents are not transferred back, essentially 'using' them for the
crafting process.

This has been tested with single full bottles, single empty/half filled
bottles, several full bottles, several half empty bottles that together
amounted to enough reagents to craft, single crafting recipes and
multiple crafting recipes at the same time, both with failures and
successes across the board, as well as stopping the crafting process mid
way through or on failure.

The only edge case I can think of that is not controlled and I do not
know how to control is if the original reagent container is destroyed
and qdeleted mid crafting recipe, which would leave the generic reagent
holder associated to it floating in nullspace. To my knowledge, this
should not break anything nor impact performance.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

It fixes an unreasonable amount of tedium and loss of reagents, as well
as an infinite crafting loop bug tied to crafting recipes using
reagents.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->