New Vile Renegade Outfit, Adept Black Light Gambesons, Skill Changes (Mostly for Vile Renegade) (#2331)
## About The Pull Request

- Removes sewing skill from thief, and zealot.
- Adepts start with black light gambesons instead of black shirts.
- Adds Tricorn, Renegade Jacket, and Renegade Collar as new clothing
items, Vile Renegade's new clothing.
![highway
man](https://github.com/user-attachments/assets/4b4eaf30-7586-4f3b-889e-3dbba9220fdb)

![RenegadeSprites](https://github.com/user-attachments/assets/e41f949a-fcc1-46e3-a034-00cfaa013eb9)

![Renegade](https://github.com/user-attachments/assets/df7fca83-6ce3-41d9-86cc-e42f26609d65)

- Gives Renegade's lockpick to thief, he had no skill, and thief
deserves it more.
- Removes Renegade's cloak
- Removes Renegade's crossbow & quiver
- Gives Renegade an iron shortsword

Skill changes
Whips 3 to 2
Swords 0 to 3
Knives 0 to 2
bows 2 to 0
crossbows 3 to 0
unarmed 1 to 2
athletics 2 to 3
swimming 4 to 2
climbing 4 to 3
riding 3 to 1
crafting 3 to 2
cooking 0 to 2
carpentry 0 to 1
reading 1 to 2
firearms 1 to 2

Stats
Perception +2 to +1
Intelligence kept at +2
Constitution -2 to -1
Strength -2 to 0
Speed 0 to 1

## Why It's Good For The Game

Vile Renegade was kind of sad to play being bad at dodging, and unable
to parry. The inquisition probably doesn't need more ranged with the
thief's bow, and the puffer if the inquisitor decides to buy one.
Levelled out their stats to be more in line with the two other adepts
now that their crossbow is gone.
Thief now has a lockpick for righteous skullduggery. 
Light Gambesons are very easy to make, I'd rather the adepts start with
black dyed ones like their shirt instead of sewing up a natural colored
one. Vile Renegade gets to keep the sewing since he's the "support" guy
now.


## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->