Verb categories reorganization (#2479)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

This PR does verb category reorganizing and renaming. The result is more
compact verb tab rows that are much more readeable and clickable
through.

**Verb category changes:**

- "Prefs - Admins" -> "Prefs" (Normal player preferences are still in
the "Options" category as before)
- "Special Verbs" -> "Special"
- "Object" category dissolved, as it had only 1 verb - point at. Living
mobs will find the verb in the "IC" category, observers in the "Spirit"
category
- "Memory" category verbs split up between "IC" and "OOC" accordingly
- Reinstated a few of the hidden "IC" category verbs - Yield, Rest and
Pray. Pray is the same as the emote version, its only also there so its
more accessible, as that emote is quite important.

Few verbs also got renamed to be more concise or accurate. For example
"Notes" became "Memories".

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Quality of life for both players and admins.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->