[BOUNTY] Adds Pestran Templar Dual-Wielding (#2483)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Gives Pestran templars the dual-wielding trait

## Why It's Good For The Game

Oversight, given they're the only templar left with multiple weapons but
no dual-wielding capabilities.

Also I eat triumphs for breakfast. 

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>