Witless Pixie flaw (#2477)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds a new flaw, **Witless Pixie** - trade a majority to the _vast_
majority of your intellect for a chance at the beautiful trait, and a
smaller chance for the ugly trait.

## Why It's Good For The Game

Ook posited it. I think it's funny. Effectively adds himbos to the game.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>