Slapcrafting improvements (#2499)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
- lots of recipes can now be started both with bundles and by flipping
the starting ingredients
- MMB the craft button to see a recipe book of basic crafts 
- leatherworking standardized so leather clothing are made from hitting
leather with needle. non-clothes items are made by hitting a rack with a
needle
- Backpack and satchels need rope (2 and 1 respectively)
- BALANCE: breaking out of ropes and chains is faster because 5 minutes
is actual CBT

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
code cleanup. qol. balance
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->