Updates Witch Hunter's Starting Gear + New Silver Whip (#2509)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

### The Outfit

<img width="217" height="200" alt="image"
src="https://github.com/user-attachments/assets/adc9672b-6bda-4390-88fb-3f3ad13695e5"
/>

Main changes here were removing the old Winter Coat, and replacing it
with Splint Armour so they actually stand a bit more of a chance when
they start taking hits. They also received a satchel because they didn't
have one before, and also the old puritan cape that once belonged to the
Inquisitor to round out the look.

### The Silver Whip

<img width="113" height="114" alt="image"
src="https://github.com/user-attachments/assets/150045cb-5bb5-440d-9746-8c140a40776a"
/>

A Whip specifically modified for dealing with Nitebeasts, and now they
start with a weapon that makes use of their Expert level in Whips &
Flails. The sprite was ported from Azure-Peak. Originally from:
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/1988

The Silver Whip can also be crafted using an Anvil. The recipe is:

- 1x Silver Bar
- 2x Cured Hide

### Skills, Stats and Traits

Skills and Stats have remained unchanged, but their Mail Training trait
has been replaced with Fast Reflexes. It seemed really weird to me that
Witch Hunter was given Mail Training and yet didn't even start with any
Medium Armour. Fast Reflexes however leans more into the idea that the
Witch Hunter is more of a duellist and should be dodging and parrying
blows instead of tanking hits.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

The Witch Hunter has always felt like it was forgotten about and left to
rot, and was overall pretty lackluster as a rare Adventurer role. This
PR aims to breathe some life into the role by giving it new starting
gear and a new weapon for them to make use of in niche cases.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->