Marriages: Fixes, feedback and secret ceremonies (#2541)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

This PR cleans up and fixes normal marriages, and adds secret marriages.

**Changes:**

- General code cleanup of the marriage code
- Better user feedback in case marriage failed
- More sensible marriage checks, like forbidding officiant to be able to
conduct their own wedding
- Apple does not burn anymore in case marriage failed
- Eora's chosen in the matchmaker event now gets an option to conduct
secret marriage outside of the church, which also won't make a world
announcement. Ceremony is otherwise the same

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Working marriages, better feedback and more eora's chosen flavor.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->