Heartfelt migrants (#2534)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Changes the heartfelt migrant wave considerably.
1. Everyone now gets varying degrees of engineering skill.
2. Adds wave downgrades. Instead of 6 players or bust, it now has 4
total versions, with only the Lord on the last wave.
3. Moved Lords leyline recharge to the magos, was probably an oversight.
The scomstone now spawns on his ring slot and he gets the heartfelt
satchel with the note containing his (optional) goal, alongside a
waterskin purifier.
4. The lady doesn't spawn with almost nothing anymore. They now get a
belt (probably oversight), pants, satchel, bow and arrow with average
skill, steel dagger and a pouch of coins.
5. Hand also starts with the scomstone on his ring slot, and only gets a
normal satchel now. If only the Lord spawns, then they should get the
note IMO.
6. The Knight now gets a full suit of steam armor, and a random amount
of steam left over in the boiler (50%-90%) as well as a minor stat
debuff and int buff, similarly to the RK steam knight.
7. Magos now gets the intrinsic leyline recharge from the lord.
8. The prior got removed from the wave, replaced by the supreme
artificer. The role itself still remains in the code for now. Felt like
a dedicated artificer made more sense to be alongside the rest of the
court, and the wave did not need any more bloat IMO.
9. The Supreme Artificer! Basicaly just a copy of the towner artificer,
but with Legendary Engineering, a purifier waterskin and the cooling
backpack.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
More heartfelty
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->